% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mietheory.R
\name{FiniteAngleCalc}
\alias{FiniteAngleCalc}
\title{Calculate Mie scattering intensity for a sphere at finite angles}
\usage{
FiniteAngleCalc(np, nm, wl, na, dmin, dmax)
}
\arguments{
\item{np}{Refractive index of the particle.}

\item{nm}{Refractive index of the medium.}

\item{wl}{Wavelength of the incident light in nanometers.}

\item{na}{Numerical aperture of the objective lens.}

\item{dmin}{Minimum diameter of the particle in microns (must be > 0).}

\item{dmax}{Maximum diameter of the particle in microns.}
}
\value{
A data frame with the following columns:
\itemize{
 \item `Diameter`: Particle diameter in microns.
 \item `AvgIpara`: Average parallel component of scattered intensity.
 \item `AvgIperp`: Average perpendicular component of scattered intensity.
 }
}
\description{
This function calculates the Mie scattering intensity for a sphere with 
given parameters across a range of particle diameters. It returns both the 
parallel and perpendicular components of the scattered light.
}
\examples{

# Refractive indices
np <- 1.6 # Beads
# n_particle <- 1.36 # lower limit for Phytoplankton
# n_particle <- 1.38 # average for Phytoplankton
# n_particle <- 1.41 # upper limit for Phytoplankton
nm <-   1.3371 # Water

# Laser wavelength (in nanometers)
wl <- 457 # (SeaFlow)       
# wl <- 488 # (InFlux)       

# Numerical aperture of the objective
na <- 0.55 # (SeaFlow)
# na <- 0.42  # (InFlux)

# Particle size range 
dmin <- 0.025   # Minimum radius (microns)
dmax <- 50      # Maximum radius (microns)

df <- FiniteAngleCalc(np, nm, wl, na, dmin, dmax)

ggplot2::ggplot(df, ggplot2::aes(x = Diameter, y = AvgIpara)) +
  ggplot2::geom_line() +
  ggplot2::scale_y_log10() +
  ggplot2::labs(y = 'Relative Intensity', x = 'Diameter (µm)') +
  ggplot2::theme_bw() 
  
# Save output
# readr::write_csv(df, paste0('mie_',round(m,3),'n.csv'))
}
