% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mietheory.R
\name{Intensity}
\alias{Intensity}
\title{Calculate scattered light intensity for a sphere}
\usage{
Intensity(m, x, ang)
}
\arguments{
\item{m}{Numeric vector of refractive index ratios. Can be a scalar.}

\item{x}{Numeric vector of size parameters. Can be a scalar.}

\item{ang}{Numeric vector of scattering angles (in radians).}
}
\value{
A list containing:
\itemize{
 \item `I1`: A matrix representing the parallel component of the scattered 
         light intensity. If `m` and/or `x` are vectors, `I1` will be a 3D 
         array with dimensions corresponding to `length(ang)`, `length(m)`, 
         and `length(x)`.
 \item `I2`: A matrix representing the perpendicular component of the scattered 
         light intensity.  If `m` and/or `x` are vectors, `I2` will be a 3D 
         array with dimensions corresponding to `length(ang)`, `length(m)`, 
         and `length(x)`.
}
}
\description{
This function calculates the scattered light intensity for a sphere, 
given its size parameter (`x`), refractive index relative to the medium (`m`), 
and scattering angle (`ang`). The incident light is assumed to be polarized.
}
\details{
If `m` and/or `x` are vectors, the output matrices `I1` and `I2` will be 3D 
arrays, with the `m` and/or `x` parameters spanning the third dimension.
}
\examples{
Intensity(m = 1.5, x = 1:10, ang = seq(0, pi, length.out = 20))
}
