% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mietheory.R
\name{RB1}
\alias{RB1}
\title{Calculate Ricatti-Bessel functions of the first kind}
\usage{
RB1(r, nmax)
}
\arguments{
\item{r}{Numeric vector of radius parameters.}

\item{nmax}{Integer specifying the maximum order of the Ricatti-Bessel function.}
}
\value{
A matrix of dimensions `nmax` by `length(r)`, where each column 
  represents the Ricatti-Bessel functions of the first kind for a 
  specific radius parameter.
}
\description{
This function calculates the Ricatti-Bessel functions of the first kind 
from order 1 to `nmax` for the given radius parameter(s) `r`, using 
downwards recursion.
}
\examples{
RB1(r = 1:10, nmax = 5)
}
\seealso{
\code{\link{RB2}} for calculating Ricatti-Bessel functions of the second kind.
}
